+++ src/constants/MoveHistory/EventTemplates/CreateUpload/proofOfServiceDocUpload.jsx
@@ -1,10 +1,11 @@
 import a from 'constants/MoveHistory/Database/Actions';
 import d from 'constants/MoveHistory/UIDisplay/DetailsTypes';
 import t from 'constants/MoveHistory/Database/Tables';
+import o from 'constants/MoveHistory/UIDisplay/Operations';
 
 export default {
   action: a.INSERT,
-  eventName: 'createUpload',
+  eventName: o.createUpload,
   tableName: t.proof_of_service_docs,
   detailsType: d.PLAIN_TEXT,
   getEventNameDisplay: () => 'Uploaded document',
